src/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.4
